+++ LongevityWorldCup.Website/Business/AthleteOgImageService.cs
@@ -265,7 +265,7 @@ public sealed class AthleteOgImageService
                 Mode = ResizeMode.Crop,
                 Position = AnchorPositionMode.Center
             }));
-            MakeCircular(profile);
+            ImageMasks.MakeCircular(profile);
 
             image.Mutate(ctx =>
             {
@@ -1081,35 +1081,6 @@ public sealed class AthleteOgImageService
         }
 
         return string.IsNullOrWhiteSpace(trimmed) ? ellipsis : trimmed + ellipsis;
-    }
-
-    private static void MakeCircular(Image<Rgba32> image)
-    {
-        var w = image.Width;
-        var h = image.Height;
-        var radius = Math.Min(w, h) / 2f;
-
-        using var mask = new Image<Rgba32>(w, h, Color.Transparent);
-        mask.Mutate(ctx =>
-        {
-            ctx.SetGraphicsOptions(new GraphicsOptions
-            {
-                Antialias = true,
-                AntialiasSubpixelDepth = 16
-            });
-            ctx.Fill(Color.White, new EllipsePolygon(w / 2f, h / 2f, radius));
-        });
-
-        image.Mutate(ctx =>
-        {
-            ctx.SetGraphicsOptions(new GraphicsOptions
-            {
-                AlphaCompositionMode = PixelAlphaCompositionMode.DestIn,
-                Antialias = true,
-                AntialiasSubpixelDepth = 16
-            });
-            ctx.DrawImage(mask, new Point(0, 0), 1f);
-        });
     }
 
     private static void DrawTextShadow(
